yulianto update

+++ database/migrations/2019_11_25_112818_create_new_pendaftares_table.php
@@ -16,15 +16,15 @@ class CreateNewPendaftaresTable extends Migration
         Schema::dropIfExists('new_pendaftares');
         Schema::create('new_pendaftares', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('pendaftar_type_id');
-            $table->integer('pendaftar_status_id');
-            $table->integer('xn1')->nullable();
-            $table->integer('xn2')->nullable();
-            $table->integer('xn3')->nullable();
-            $table->integer('xn4')->nullable();
-            $table->integer('xn5')->nullable();
-            $table->integer('xn6')->nullable();
-            $table->integer('xn7')->nullable();
+            $table->bigInteger('pendaftar_type_id');
+            $table->bigInteger('pendaftar_status_id');
+            $table->bigInteger('xn1')->nullable();
+            $table->bigInteger('xn2')->nullable();
+            $table->bigInteger('xn3')->nullable();
+            $table->bigInteger('xn4')->nullable();
+            $table->bigInteger('xn5')->nullable();
+            $table->bigInteger('xn6')->nullable();
+            $table->bigInteger('xn7')->nullable();
             $table->text('xs1')->nullable();
             $table->text('xs2')->nullable();
             $table->text('xs3')->nullable();
